FoodSpawner
randomly picks a child of the game object "Mid" which is the middle area and spawn "food" above it. It has 2 variables one for maximum amount of food on the map and one for current amount of food on the map. Those variables are for when player collides with food another one will be spawned.